Remove home-made driver refcounting
By setting the right module in the .owner of the file operations
(within dewb_sysfs.c), we already allow the kernel to properly do his own
refcounting over the module, and thus we should not bother about it.
Also, it seems that trying to do it ourselves is risky, since the code
executed is already part of the module itself.

Another bonus: It is less error-prone not to manage it ourselves.