Get rid of frame_status_ok() in header_get_uint
The frame_status_ok function was written to check for a HTTP/1.1 200 OK
response from the server, but some of our requests sometimes get a 200-class
response which is not the code 200. This led to some issues when issuing
getrange commands: the 206 partial response prevented the function which
checked whether the response was read fully to do its job properly when the
response was not exactly a 200 OK.

This is now fixed by removing the frame_status_ok call, and checking the HTTP
Response Status beforehand wherever the header_get_uint function is called and
the status is relevant.