First part of the work for #60 (+ finalize #59)
 - Change the way to manage disconnects while sending a request or receiving a
response.
 - Now Only retry once in case of disconnect

The only unadressed part of #59 was the return status of the operation. For
simplicity, either the response could be fully retrieved, and we return the
length of the reponse; or it couldn't and we return the last error code
(whichever it is).

Thus we will consider that this patch completes the fix to issue #59.